+++ Sources/Base/OAuth2KeychainAccount.swift
@@ -5,14 +5,28 @@
 //  Created by David Kraus on 09/03/16.
 //  Copyright © 2016 Pascal Pfiffner. All rights reserved.
 //
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
 
 import Foundation
-#if IMPORT_SWIFT_KEYCHAIN       // experimental for SwiftKeychain integration via CocoaPods (iOS only)
-    import SwiftKeychain
-#elseif !NO_KEYCHAIN_IMPORT     // needs to be imported when using `swift build`, not when building via Xcode
-    import SwiftKeychain
+#if !NO_KEYCHAIN_IMPORT     // needs to be imported when using `swift build` or with CocoaPods, not when building via Xcode
+import SwiftKeychain
 #endif
 
+
+/**
+Base keychain integration handler for OAuth2.
+*/
 struct OAuth2KeychainAccount: KeychainGenericPasswordType {
     let accountName: String
     let internalServiceName: String
@@ -33,3 +47,4 @@ struct OAuth2KeychainAccount: KeychainGenericPasswordType {
         self.data = data
     }
 }
+